Upgraded the garpage collector, instead of using for each enviroment their last_reference, which is local, I use now global reference counting

+++ src/interpreter/variable_slot.rs
@@ -3,5 +3,6 @@ use super::value::Value;
 #[derive(Debug, Clone)]
 pub enum VariableSlot {
     Variable(Value),
-    Reference(u128)
+    Reference(u64),
+    // ListSlot(Value)
 }
